refactored code action code a bit
