+++ server/src/server.ts
@@ -4,12 +4,23 @@
  * Implements pyright language server.
  */
 
+import * as fs from 'fs';
+import * as path from 'path';
 import { isArray } from 'util';
+import { CodeAction, CodeActionParams, Command, ExecuteCommandParams } from 'vscode-languageserver';
+import { CommandController } from './commands/commandController';
+import * as debug from './common/debug';
+import { convertUriToPath, getDirectoryPath, normalizeSlashes } from './common/pathUtils';
 import { LanguageServerBase, ServerSettings, WorkspaceServiceInstance } from './languageServerBase';
+import { CodeActionProvider } from './languageService/codeActionProvider';
 
 class Server extends LanguageServerBase {
+    private _controller: CommandController;
+
     constructor() {
-        super('Pyright');
+        debug.assert(fs.existsSync(path.join(__dirname, 'typeshed-fallback')), 'Unable to locate typeshed fallback folder.');
+        super('Pyright', getDirectoryPath(__dirname));
+        this._controller = new CommandController(this);
     }
 
     async getSettings(workspace: WorkspaceServiceInstance): Promise<ServerSettings> {
@@ -17,15 +28,15 @@ class Server extends LanguageServerBase {
         try {
             const pythonSection = await this.getConfiguration(workspace, 'python');
             if (pythonSection) {
-                serverSettings.pythonPath = pythonSection.pythonPath;
-                serverSettings.venvPath = pythonSection.venvPath;
+                serverSettings.pythonPath = normalizeSlashes(pythonSection.pythonPath);
+                serverSettings.venvPath = normalizeSlashes(pythonSection.venvPath);
             }
 
             const pythonAnalysisSection = await this.getConfiguration(workspace, 'python.analysis');
             if (pythonAnalysisSection) {
                 const typeshedPaths = pythonAnalysisSection.typeshedPaths;
                 if (typeshedPaths && isArray(typeshedPaths) && typeshedPaths.length > 0) {
-                    serverSettings.typeshedPath = typeshedPaths[0];
+                    serverSettings.typeshedPath = normalizeSlashes(typeshedPaths[0]);
                 }
             }
 
@@ -40,9 +51,21 @@ class Server extends LanguageServerBase {
                 serverSettings.disableLanguageServices = false;
             }
         } catch (error) {
-            this.console.log(`Error reading settings: ${error}`);
+            this.console.log(`Error reading settings: ${ error }`);
         }
         return serverSettings;
+    }
+
+    protected executeCommand(cmdParams: ExecuteCommandParams): Promise<any> {
+        return this._controller.execute(cmdParams);
+    }
+
+    protected async executeCodeAction(params: CodeActionParams): Promise<(Command | CodeAction)[] | undefined | null> {
+        this.recordUserInteractionTime();
+
+        const filePath = convertUriToPath(params.textDocument.uri);
+        const workspace = this.workspaceMap.getWorkspaceForFile(filePath);
+        return CodeActionProvider.getCodeActionsForPosition(workspace, filePath, params.range);
     }
 }
